add: `tablec_full`
This commit adds `tablec_full` function, allowing to check if the hashtable is full, and if not, how many spaces left.

+++ tablec.h
@@ -22,6 +22,6 @@ void tablec_del(struct tablec_ht *tablec, char *key);
 
 struct tablec_bucket tablec_get(struct tablec_ht *tablec, char *key);
 
-int tablec_full(struct tablec_ht *tablec);
+size_t tablec_full(struct tablec_ht *tablec);
 
 #endif /* TABLEC_H */
